fix(compress): reject invalid compressed replacements
